+++ packages/orq-rc/src/models/operations/updatedataset.ts
@@ -182,7 +182,7 @@ export const UpdateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-05-20T07:46:35.676Z",
+    "2026-05-20T13:21:07.678Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {